Refactor: consolidate imports from logic module in CloudSyncContext and SettingsPage
- Updated CloudSyncContext to import computeDataHash and isLogoutInProgress from the logic module.
- Streamlined imports in SettingsPage by consolidating utility functions from logic.
